# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Miva Campus Maintenance API",
    "description": "RESTful API for the Miva Open University Campus Maintenance Complaint & Service Request Platform. Provides CRUD operations for service requests, user management, task assignment, status logging, and data export.",
    "version": "1.0.0",
    "contact": {
      "name": "Miva Open University — Facilities Management",
      "email": "facilities@miva.edu.ng"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://mivaca.vercel.app/api",
      "description": "Production Server (Vercel)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from auth session"
      }
    },
    "schemas": {
      "ServiceRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "requester_id": { "type": "string", "format": "uuid" },
          "category_id": { "type": "integer" },
          "title": { "type": "string", "maxLength": 150 },
          "description": { "type": "string", "maxLength": 2000 },
          "location": { "type": "string" },
          "priority": { "type": "string", "enum": ["low", "medium", "high", "critical"] },
          "status": { "type": "string", "enum": ["pending", "assigned", "in_progress", "completed", "cancelled"] },
          "evidence_urls": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateRequestBody": {
        "type": "object",
        "required": ["title", "description", "category_id", "location", "priority"],
        "properties": {
          "title": { "type": "string", "minLength": 5, "maxLength": 150 },
          "description": { "type": "string", "minLength": 20, "maxLength": 2000 },
          "category_id": { "type": "integer" },
          "location": { "type": "string", "minLength": 3 },
          "priority": { "type": "string", "enum": ["low", "medium", "high", "critical"] },
          "evidence_urls": { "type": "array", "items": { "type": "string", "format": "uri" }, "maxItems": 5 }
        }
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "full_name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role_id": { "type": "integer" },
          "department": { "type": "string", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "is_active": { "type": "boolean" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "PaginatedRequests": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ServiceRequest" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
    "/requests": {
      "get": {
        "tags": ["Service Requests"],
        "summary": "List service requests",
        "description": "Returns a paginated list of service requests filtered by the caller's role. Students see only their own; officers see assigned; admins see all.",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["all", "pending", "assigned", "in_progress", "completed", "cancelled"] } },
          { "name": "category", "in": "query", "schema": { "type": "integer" } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["all", "low", "medium", "high", "critical"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search in title, description, and location" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "updated_at", "priority"], "default": "created_at" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedRequests" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Service Requests"],
        "summary": "Create a new service request",
        "description": "Submits a new maintenance request. Automatically triggers a confirmation email to the requester.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateRequestBody" } } }
        },
        "responses": {
          "201": { "description": "Created successfully", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/ServiceRequest" } } } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List maintenance officers (Admin only)",
        "description": "Returns all active maintenance officer profiles for use in the assignment dropdown.",
        "responses": {
          "200": { "description": "List of officers", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Profile" } } } } } } },
          "403": { "description": "Forbidden — Admin only" }
        }
      }
    },
...
  }
}