# src/app/api/requests/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { createClient } from '@/lib/supabase/server';
import { adminClient } from '@/lib/supabase/admin';
import { createRequestSchema } from '@/lib/validations/request.schema';
import { dispatchNotification } from '@/lib/actions/notification.actions';
import type { RequestFilters } from '@/types';

// -----------------------------------------------
// GET /api/requests — List requests (role-filtered)
// -----------------------------------------------
export async function GET(request: NextRequest) {
  try {
    const supabase = await createClient();
    const { data: { user }, error: authError } = await supabase.auth.getUser();
    if (authError || !user) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
    }

    const { data: profile } = await supabase
      .from('profiles')
      .select('role_id')
      .eq('id', user.id)
      .single();
    
    const ROLE_MAP: Record<number, string> = { 1: 'student', 2: 'maintenance_officer', 3: 'admin' };
    const role = ROLE_MAP[profile?.role_id as number] ?? 'student';

    const { searchParams } = new URL(request.url);
    const categoryParam = searchParams.get('category');
    const filters: RequestFilters = {
      status: (searchParams.get('status') as RequestFilters['status']) ?? 'all',
      category: (categoryParam && categoryParam !== 'all') ? Number(categoryParam) : 'all',
      priority: (searchParams.get('priority') as RequestFilters['priority']) ?? 'all',
      search: searchParams.get('search') ?? '',
      page: Number(searchParams.get('page') ?? 1),
      limit: Number(searchParams.get('limit') ?? 20),
      sort: (searchParams.get('sort') as RequestFilters['sort']) ?? 'created_at',
      order: (searchParams.get('order') as 'asc' | 'desc') ?? 'desc',
    };

    const offset = (filters.page! - 1) * filters.limit!;

    let query = adminClient
      .from('service_requests')
      .select(`
        *,
        profiles!service_requests_requester_id_fkey(id, full_name, email, department),
        request_categories(id, name, icon),
        assignments(id, officer_id, assigned_at, notes, profiles!assignments_officer_id_fkey(full_name, email))
      `, { count: 'exact' });

    // Role-based scoping
    if (role === 'student') {
      query = query.eq('requester_id', user.id);
    } else if (role === 'maintenance_officer') {
      const { data: myAssignments } = await adminClient
        .from('assignments')
        .select('request_id')
        .eq('officer_id', user.id);
      const ids = (myAssignments ?? []).map((a) => a.request_id);
      if (ids.length === 0) return NextResponse.json({ data: [], total: 0, page: 1, limit: filters.limit, totalPages: 0 });
      query = query.in('id', ids);
    }
    // admin sees all

    // Apply filters
    const userId = searchParams.get('userId');
    if (role === 'admin' && userId) query = query.eq('requester_id', userId);
    
    if (filters.status && filters.status !== 'all') query = query.eq('status', filters.status);
    if (filters.category && filters.category !== 'all') query = query.eq('category_id', filters.category);
    if (filters.priority && filters.priority !== 'all') query = query.eq('priority', filters.priority);
    if (filters.search) {
      query = query.or(`title.ilike.%${filters.search}%,description.ilike.%${filters.search}%,location.ilike.%${filters.search}%`);
    }

    query = query
      .order(filters.sort!, { ascending: filters.order === 'asc' })
      .range(offset, offset + filters.limit! - 1);

    const { data, error, count } = await query;
    if (error) return NextResponse.json({ error: error.message }, { status: 500 });

    const total = count ?? 0;
    const totalPages = Math.ceil(total / filters.limit!);

    return NextResponse.json({ data, total, page: filters.page, limit: filters.limit, totalPages });
  } catch (err) {
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
  }
}

// -----------------------------------------------
// POST /api/requests — Create new service request
// -----------------------------------------------
export async function POST(request: NextRequest) {
  try {
    const supabase = await createClient();
    const { data: { user }, error: authError } = await supabase.auth.getUser();
    if (authError || !user) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
    }

    const body = await request.json();
    const result = createRequestSchema.safeParse(body);
    if (!result.success) {
      return NextResponse.json({ error: result.error.issues[0].message }, { status: 400 });
    }

    const { data, error } = await adminClient
      .from('service_requests')
      .insert({ ...result.data, requester_id: user.id })
      .select()
      .single();

    if (error) return NextResponse.json({ error: error.message }, { status: 500 });

    // Log initial status
    await adminClient.from('status_logs').insert({
      request_id: data.id,
      changed_by: user.id,
      old_status: null,
      new_status: 'pending',
      remarks: 'Request submitted',
    });

    // Trigger notification for Student
    dispatchNotification({
      userId: user.id,
      title: 'Request Submitted',
      message: `Your request "${data.title}" has been received.`,
      type: 'request_created',
      link: `/student/requests`,
      sendEmail: true,
      requestId: data.id,
    });

    // Trigger notification for Admins
    const { data: admins } = await adminClient.from('profiles').select('id').eq('role_id', 3);
    admins?.forEach(admin => {
      dispatchNotification({
        userId: admin.id,
        title: 'New Request Received',
        message: `A new request "${data.title}" was submitted.`,
        type: 'request_created',
        link: `/admin/requests`,
      });
    });

    return NextResponse.json({ data }, { status: 201 });
  } catch (err) {
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
  }
}